Fix page rendering on the server
The root cause was the use of local state for queue, which can't be set in asyncData(). To circumvent this, fetchData() was called again during create() to initialize the queue on the client side. However, since fetchData() is executing asynchronous, the call had no effect on the server side. Thus the v-if="queue" hindered the page from rendering.

The solution is to access the queue directly through the store using a computed property instead of a local state variable. The call to fetchData() could subsequently be removed, and is now only used to refresh the store when a route change occurs.